requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.13.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
asttokens==3.0.0
astunparse==1.6.3
atomicwrites==1.4.1
attrs==24.2.0
bleach==6.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
coverage==7.6.8
cycler==0.12.1
daal==2025.5.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.5.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
distro==1.9.0
executing==2.2.0
filelock==3.17.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-Reuploaded==1.2.0
Flask-SQLAlchemy==3.0.1
Flask-Uploads==0.2.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
geckodriver-autoinstaller==0.1.0
gevent==24.10.3
geventhttpclient==2.3.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.29.2
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.2.6
kaggle==1.6.17
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
locust==2.20.1
Mako==1.3.5
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
namex==0.0.9
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.1.3
openai==1.51.2
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
progressbar2==4.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
py==1.11.0
py4j==0.10.9.7
pyarrow==19.0.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
pyspark==3.5.0
pytest==7.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-utils==3.9.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==14.0.0
roundrobin==0.0.4
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.5.1
scikit-learn-intelex==2025.5.0
scipy==1.15.2
seaborn==0.13.2
selenium==4.26.1
setuptools==75.1.0
shap==0.47.2
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.18
stack-data==0.6.3
streamlit==1.45.0
sympy==1.14.0
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.3.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
torch==2.7.0
torch-geometric==2.6.1
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.49.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.1.2
xgboost==3.0.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.1.1
